% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apipkgen-package.R
\docType{package}
\name{apipkgen-package}
\alias{apipkgen-package}
\alias{apipkgen}
\title{apipkgen}
\description{
apipkgen
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
