% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg.R
\name{generate_pkg}
\alias{generate_pkg}
\title{HTTP API package generator}
\usage{
generate_pkg(
  pkg_path,
  template_path = NULL,
  http_lib = "crul",
  base_url = NULL
)
}
\arguments{
\item{pkg_path}{(character) Path to the new package, where last part will be
the package name}

\item{template_path}{(character) path to your yaml template file. by default,
we use a demo template yaml file}

\item{http_lib}{(character) one of \code{crul} (default) or \code{httr}}

\item{base_url}{(character) Base URL. If \code{NULL}, defaults to \code{baseUrl} or is
formed from \code{schemes}, \code{host}, \code{basePath} as specified in the template
file.}
}
\description{
HTTP API package generator
}
\examples{
\dontrun{
generate_pkg(pkg_path = "mypkg")

# from a Swagger spec
url <- "https://raw.githubusercontent.com/ropenscilabs/apispecs/master/swagger/crossref.yml"
# x <- paste0(readLines(url), collapse = "\n")
# yaml::yaml.load(string = x)
download.file(url, "crossref.yml")
my_dir <- file.path(tempdir(), "apples")
generate_pkg(pkg_path = my_dir, template_path = "crossref.yml")
}
}
