% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_constants.R
\name{write_constants}
\alias{write_constants}
\title{Constants to zzz}
\usage{
write_constants(template_path = NULL, outfile = "zzz.R",
  base_url = NULL)
}
\arguments{
\item{template_path}{(character) path to your yaml template file. by default,
we use a demo template yaml file}

\item{outfile}{(character) Path to write your functions to}

\item{base_url}{(character) optional. if not given, use base url from spec.
you can give it here and it will override anything in the spec}
}
\value{
returns silently, after writing things to zzz.R
}
\description{
Constants to zzz
}
