% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fxns.R
\name{write_fxns}
\alias{write_fxns}
\title{Function generator}
\usage{
write_fxns(template_path = NULL, outfile = "http-fxns.R")
}
\arguments{
\item{template_path}{(character) path to your yaml template file. by default,
we use a demo template yaml file}

\item{outfile}{(character) Path to write your functions to}
}
\value{
returns silently, after writing all functions into a single file,
given by \code{outfile}
}
\description{
Function generator
}
\examples{
\dontrun{
write_fxns()
}
}

