% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_helpers.R
\name{write_helpers}
\alias{write_helpers}
\title{Write http helpers}
\usage{
write_helpers(outfile = "http-helpers.R", http_lib = "crul")
}
\arguments{
\item{outfile}{(character) Path to write http helpers file to}

\item{http_lib}{(character) one of \code{crul} (default) or \code{httr}}
}
\description{
Write http helpers
}
\examples{
\dontrun{
write_helpers()
}
}
