% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apipkgen-package.R
\docType{package}
\name{apipkgen-package}
\alias{apipkgen}
\alias{apipkgen-package}
\title{apipkgen: Package Generator for HTTP API Wrapper Packages}
\description{
Package Generator for HTTP API Wrapper Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sckott/apipkgen}
  \item \url{http://sckott.github.io/apipkgen/}
  \item Report bugs at \url{https://github.com/sckott/apipkgen/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com}

}
\keyword{internal}
