% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_helpers.R
\name{write_helpers}
\alias{write_helpers}
\title{Write http helpers}
\usage{
write_helpers(outfile = "http-helpers.R")
}
\arguments{
\item{outfile}{(character) Path to write http helpers file to}
}
\description{
Write http helpers
}
\examples{
\dontrun{
write_helpers()
}
}

